double precision

+++ src/cuda.h
@@ -18,7 +18,7 @@
 int cuda_init(struct finite_element_problem *restrict p);
 void cuda_destroy(struct finite_element_problem *restrict p);
 
-int gpu_conj_gradient(struct finite_element_problem *restrict p, float tolerance);
+int gpu_conj_gradient(struct finite_element_problem *restrict p, double tolerance);
 
 #endif /* GPU_COMPUTE */
